Rename `drivers/arrival` to `drivers/common`
